Fix Studio files in VirtualFileExplorer module.
