+++ modules/virtual-file-explorer/.abpstudio/state.json
@@ -1,0 +1,11 @@
+﻿{
+  "selectedKubernetesProfile": null,
+  "solutionRunner": {
+    "selectedProfile": null,
+    "targetFrameworks": [],
+    "applicationsStartingWithoutBuild": [],
+    "applicationsWithoutAutoRefreshBrowserOnRestart": [],
+    "applicationBatchStartStates": [],
+    "folderBatchStartStates": []
+  }
+}
