+++ modules/virtual-file-explorer/Volo.Abp.VirtualFileExplorer.abpsln
@@ -3,5 +3,6 @@
     "Volo.Abp.VirtualFileExplorer": {
       "path": "Volo.Abp.VirtualFileExplorer.abpmdl"
     }
-  }
+  },
+  "id": "d900c5b2-d907-4abb-bc38-b72f80006fe0"
 }
