+++ modules/virtual-file-explorer/app/Data/DemoAppDbContext.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Volo.Abp.EntityFrameworkCore;
+using Volo.Abp.Identity.EntityFrameworkCore;
+using Volo.Abp.PermissionManagement.EntityFrameworkCore;
+
+namespace DemoApp.Data;
+
+public class DemoAppDbContext : AbpDbContext<DemoAppDbContext>
+{
+    public DemoAppDbContext(DbContextOptions<DemoAppDbContext> options)
+        : base(options)
+    {
+    }
+
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        base.OnModelCreating(builder);
+
+        builder.ConfigurePermissionManagement();
+        builder.ConfigureIdentity();
+    }
+}
