--- modules/virtual-file-explorer/app/Data/DempAppDbContext.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Volo.Abp.EntityFrameworkCore;
-using Volo.Abp.Identity.EntityFrameworkCore;
-using Volo.Abp.PermissionManagement.EntityFrameworkCore;
-
-namespace DempApp.Data;
-
-public class DempAppDbContext : AbpDbContext<DempAppDbContext>
-{
-    public DempAppDbContext(DbContextOptions<DempAppDbContext> options)
-        : base(options)
-    {
-    }
-
-    protected override void OnModelCreating(ModelBuilder builder)
-    {
-        base.OnModelCreating(builder);
-
-        builder.ConfigurePermissionManagement();
-        builder.ConfigureIdentity();
-    }
-}
