+++ modules/virtual-file-explorer/app/DemoApp.abppkg
@@ -1,0 +1,3 @@
+{
+  "role": "host.mvc"
+}
