+++ modules/virtual-file-explorer/app/DemoApp.csproj
