+++ modules/virtual-file-explorer/app/Pages/Index.cshtml
@@ -1,4 +1,4 @@
 ﻿@page
 @using Volo.Abp.Users
-@model DempApp.Pages.IndexModel
+@model DemoApp.Pages.IndexModel
 @inject ICurrentUser CurrentUser
