+++ modules/virtual-file-explorer/app/Properties/launchSettings.json
@@ -1,4 +1,4 @@
-﻿{
+{
   "iisSettings": {
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
@@ -15,7 +15,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "DempApp": {
+    "DemoApp": {
       "commandName": "Project",
       "launchBrowser": true,
       "environmentVariables": {
