+++ modules/virtual-file-explorer/app/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "Default": "Server=localhost;Database=VirtualFileExplorerDempApp;Trusted_Connection=True;TrustServerCertificate=True"
+    "Default": "Server=(LocalDb)\\MSSQLLocalDB;Database=VirtualFileExplorerDemoApp;Trusted_Connection=True;TrustServerCertificate=True"
   },
   "Settings": {
     "Abp.Identity.Password.RequireNonAlphanumeric": "false",
